Added enrollment page with validations, dynamic addition, image loading, fade-in effect, and responsiveness. Used HTML5, CSS3, JavaScript, Bootstrap.
his is an enrollment page created using HTML5, CSS3, and JavaScript. It features necessary validations and error messages for all form fields. Upon form submission, the enrolled student is dynamically added to a table using JavaScript. The image is loaded from the submitted URL, and the website link opens in a new tab upon clicking. The new enrollment is added with a fade-in effect achieved using CSS animations. The webpage is responsive and designed to adapt to different resolutions and screen sizes. The Bootstrap framework is used for layout and forms.